move AdminContent directory and extract OrgContext
