+++ frontend/src/pages/admin/OrgContext.js
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export const OrgContext = createContext(null);
